refactor: extract SignalingModule and CallState from part-of library boundary
- call_state.dart is now a standalone library with its own freezed file
- signaling_module.dart is now a standalone importable library
- _SignalingHandlers extension moved to new signaling_handlers.dart (part of call_bloc)
- attachDelegate() replaces direct _delegate field access
- call_bloc.dart re-exports both for backward compatibility
